few improvements
this version is particularly weird. when i remove the code to setup the bltr arrays in Util.c's function util_setup(), my conversion tests work, when i don't the tests fail. i looked at the code that was being removed carefully, it does not reference any of the tlbr arrays! at initialization, the arrays are not indicated to be the same (separate malloc calls). my only thought is that somehow the compiler is not noticing the difference between the array names!